Striae worker and export encryption modularization
