+++ package.json
@@ -88,11 +88,13 @@
     "typecheck": "react-router typegen && tsc",
     "typegen": "wrangler types",
     "preview": "npm run build && wrangler pages dev",
-    "test": "npm run test:app && npm run test:workers:data && npm run test:workers:audit && npm run test:workers:user",
+    "test": "npm run test:app && npm run test:workers:data && npm run test:workers:audit && npm run test:workers:user && npm run test:workers:image && npm run test:workers:files",
     "test:app": "vitest run --config tests/app/vitest.config.ts",
     "test:workers:data": "vitest run --config tests/workers/data/vitest.config.mjs",
     "test:workers:audit": "vitest run --config tests/workers/audit/vitest.config.mjs",
     "test:workers:user": "vitest run --config tests/workers/user/vitest.config.mjs",
+    "test:workers:image": "vitest run --config tests/workers/image/vitest.config.mjs",
+    "test:workers:files": "vitest run --config tests/workers/files/vitest.config.mjs",
     "test:watch": "vitest --config tests/app/vitest.config.ts",
     "test:coverage": "vitest run --config tests/app/vitest.config.ts --coverage",
     "delete-account": "node ./scripts/delete-account.mjs",
