+++ workers/image-worker/src/types.ts
@@ -17,12 +17,7 @@ export interface KeyRegistryPayload {
 	keys?: unknown;
 }
 
-export interface PrivateKeyRegistry {
-	activeKeyId: string | null;
-	keys: Record<string, string>;
-}
-
-export type DecryptionTelemetryOutcome = 'primary-hit' | 'fallback-hit' | 'all-failed';
+export type { PrivateKeyRegistry, DecryptionTelemetryOutcome } from '../../../shared/registry/key-candidates';
 
 export interface UploadResult {
 	id: string;
